Change of command
INCLUDE .. with context is hard to parse with regexes, because there's a
string in the include statement including a filename and we can't easily
regex-parse something that requires a lexer and parser to properly
decode.  Instead, change the command itself.

This also simplifies the regex, and there's a reworking of the other
code in there as well since that has a number of conditions that
should not need checking and have no tests ('matched string is false').